feat(client.reactNative): implement automatic language selection using geo-location and remove AutoLanguageSelector component
