+++ reeltok.client/reeltok.client.reactNative/src/hooks/useGeoLocation.ts
@@ -1,9 +1,14 @@
 import { useEffect, useState } from 'react'
 import * as Location from 'expo-location'
+import { useAppDispatch } from '../hooks/useAppDispatch'
+import { useAppSelector } from '../hooks/useAppSelector'
+import { changeLanguageThunk } from '../redux/thunks/settingsThunks'
 
-export const useGeoLocation = () => {
+const useGeoLocationUtils = () => {
   const [location, setLocation] = useState<Location.LocationObjectCoords | null>(null)
   const [country, setCountry] = useState<string>('')
+  const dispatch = useAppDispatch()
+  const language = useAppSelector((state) => state.settings.language)
 
   useEffect(() => {
     ;(async () => {
@@ -13,21 +18,33 @@ export const useGeoLocation = () => {
         return
       }
 
-      // Get user's location
       let locationData = await Location.getCurrentPositionAsync({})
       setLocation(locationData.coords)
 
-      // Reverse geocode to get country
       let reverseGeocode = await Location.reverseGeocodeAsync({
         latitude: locationData.coords.latitude,
         longitude: locationData.coords.longitude,
       })
 
       if (reverseGeocode.length > 0) {
-        setCountry(reverseGeocode[0].country || '') // Get country name
+        setCountry(reverseGeocode[0].country || '')
       }
     })()
   }, [])
 
+  useEffect(() => {
+    if (!language || !language.locale) {
+      const selectedLanguage = country === 'Denmark' ? 'da_DK' : 'en_GB'
+      dispatch(
+        changeLanguageThunk({
+          label: selectedLanguage === 'da_DK' ? 'Danish' : 'English',
+          value: selectedLanguage,
+        })
+      )
+    }
+  }, [country, language, dispatch])
+
   return { location, country }
 }
+
+export default useGeoLocationUtils
